Partially reverted "zink.apple: Collection of fixes to allow Zink to
startup and partially run on macOS"
(https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/28259)

+++ src/freedreno/vulkan/tu_image.h
@@ -23,7 +23,7 @@ struct tu_image
    struct vk_image vk;
 
    struct fdl_layout layout[3];
-   uint32_t total_size;
+   uint64_t total_size;
 
 #if DETECT_OS_ANDROID
    /* For VK_ANDROID_native_buffer, the WSI image owns the memory, */
